web_application/backend/requirements.txt
annotated-types==0.7.0
anyio==4.7.0
asgiref==3.8.1
boto3==1.35.76
botocore==1.35.76
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
exceptiongroup==1.2.2
fastapi==0.115.6
greenlet==3.1.1
h11==0.14.0
idna==3.10
jmespath==1.0.1
psycopg==3.2.3
psycopg2-binary==2.9.10
pydantic==2.10.3
pydantic_core==2.27.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.19
requests==2.32.3
s3transfer==0.10.4
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.36
starlette==0.41.3
typing_extensions==4.12.2
urllib3==2.2.3
uvicorn==0.32.1
